model_vlm/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None '.\requirements.in'
#
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
    #   uvicorn
fastapi==0.117.1
    # via -r requirements.in
filelock==3.19.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2025.9.0
    # via
    #   huggingface-hub
    #   torch
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via ollama
huggingface-hub==0.35.1
    # via
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via torch
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via torch
numpy==2.3.3
    # via
    #   torchvision
    #   transformers
ollama==0.5.4
    # via -r requirements.in
packaging==25.0
    # via
    #   huggingface-hub
    #   transformers
pillow==11.3.0
    # via
    #   -r requirements.in
    #   torchvision
pydantic==2.11.9
    # via
    #   -r requirements.in
    #   fastapi
    #   ollama
pydantic-core==2.33.2
    # via pydantic
python-dotenv==1.1.1
    # via uvicorn
python-multipart==0.0.20
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   transformers
    #   uvicorn
regex==2025.9.18
    # via transformers
requests==2.32.5
    # via
    #   huggingface-hub
    #   transformers
safetensors==0.6.2
    # via transformers
sniffio==1.3.1
    # via anyio
starlette==0.48.0
    # via fastapi
sympy==1.14.0
    # via torch
tokenizers==0.22.1
    # via transformers
torch==2.8.0
    # via
    #   -r requirements.in
    #   torchvision
torchvision==0.23.0
    # via -r requirements.in
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   transformers
transformers==4.56.2
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   torch
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via requests
uvicorn[standard]==0.37.0
    # via -r requirements.in
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn

# The following packages are considered to be unsafe in a requirements file:
# setuptools
